Reorganize datamovement_controller
Allow the bulk of datamovement_controller to be usable from a forthcoming
copy-offload daemon.
